requirements.txt
alembic==1.3.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-JSON==0.3.4
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
protobuf==3.6.1
python-crontab==2.4.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
